vidder_handlers/user_vidder.py
"""
👥 VidderTech User Management Handlers
Built by VidderTech - The Future of Quiz Bots
"""

import logging
from telegram import Update
from telegram.ext import ContextTypes, CommandHandler
from telegram.constants import ParseMode

logger = logging.getLogger('vidder.handlers.user')

class VidderUserHandlers:
    """👥 VidderTech User Management"""
    
    @staticmethod
    async def add_command(update: Update, context: ContextTypes.DEFAULT_TYPE):
        """👥 /add - Add User to Paid Quiz"""
        await update.message.reply_text("👥 **VidderTech User Management**\n\n🚧 Coming soon!", parse_mode=ParseMode.MARKDOWN)
    
    @staticmethod
    async def rem_command(update: Update, context: ContextTypes.DEFAULT_TYPE):
        """👥 /rem - Remove User Access"""
        await update.message.reply_text("👥 **VidderTech User Removal**\n\n🚧 Coming soon!", parse_mode=ParseMode.MARKDOWN)

def register_user_vidder_handlers(app) -> int:
    """Register user management handlers"""
    handlers = [
        CommandHandler("add", VidderUserHandlers.add_command),
        CommandHandler("rem", VidderUserHandlers.rem_command)
    ]
    for handler in handlers:
        app.add_handler(handler)
    return len(handlers)