requirements.txt
# 🚀 VidderTech Advanced Quiz Bot - Dependencies
# Built by VidderTech - The Future of Quiz Bots

# Core Telegram Bot Framework
python-telegram-bot[webhooks]==20.7
python-telegram-bot==20.7

# Database Operations  
aiosqlite==0.19.0
sqlalchemy==2.0.25

# HTTP Requests & APIs
requests==2.31.0
aiohttp==3.9.1
httpx==0.25.2

# Web Scraping & Content Extraction
beautifulsoup4==4.12.2
lxml==4.9.4
selenium==4.16.2
scrapy==2.11.0

# OCR & Image Processing
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78
pdf2image==1.16.3
PyPDF2==3.0.1

# AI & Machine Learning
openai==1.6.1
google-generativeai==0.3.2
transformers==4.36.2
torch==2.1.2
scikit-learn==1.3.2

# Text Processing & NLP
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1
langdetect==1.0.9

# Template Engine & Reporting
Jinja2==3.1.2
weasyprint==60.2
reportlab==4.0.8
matplotlib==3.8.2
plotly==5.17.0

# Configuration & Environment
python-dotenv==1.0.0
pydantic==2.5.2
pyyaml==6.0.1
toml==0.10.2

# Date & Time Processing
python-dateutil==2.8.2
pytz==2023.3
arrow==1.3.0

# Async & Concurrency
asyncio
asyncpg==0.29.0
aioredis==2.0.1
celery==5.3.4

# Security & Cryptography
cryptography==41.0.8
passlib==1.7.4
jwt==1.3.1
bcrypt==4.1.2

# File Processing & Storage
cloudinary==1.36.0
boto3==1.34.12
azure-storage-blob==12.19.0

# Caching & Performance
redis==5.0.1
memcached==1.59
cachetools==5.3.2

# Monitoring & Logging
loguru==0.7.2
sentry-sdk==1.39.2
prometheus-client==0.19.0

# Data Validation & Serialization
marshmallow==3.20.2
cerberus==1.3.4
jsonschema==4.20.0

# Utilities
uuid
hashlib
json5==0.9.14
python-magic==0.4.27
chardet==5.2.0
tqdm==4.66.1

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
mypy==1.8.0

# Additional Integrations
google-api-python-client==2.112.0
youtube-dl==2021.12.17
wikipedia==1.4.0
feedparser==6.0.10

# Process Management
setproctitle==1.3.3
psutil==5.9.6

# Internationalization
babel==2.14.0
gettext==4.1.0

# Enhanced Features
qrcode==7.4.2
python-barcode==0.15.1
fpdf2==2.7.6

# Total: 80+ carefully selected packages for enterprise functionality